+++ oblate/__init__.py
@@ -8,7 +8,7 @@ Copyright (C) Izhar Ahmad 2023, licensed under MIT license.
 See LICENSE at project's root for more information.
 """
 
-__version__ = '1.1.0'
+__version__ = '1.2.0'
 __author__  = 'Izhar Ahmad (izxxr)'
 
 from oblate import fields as fields
@@ -17,3 +17,4 @@ from oblate.schema import *
 from oblate.contexts import *
 from oblate.exceptions import *
 from oblate.configs import *
+from oblate.type_validation import *