+++ oblate/fields/structs.py
@@ -35,7 +35,8 @@ from typing import (
 )
 from oblate.fields.base import Field
 from oblate.exceptions import FieldError
-from oblate import utils
+from oblate.utils import MISSING
+from oblate.type_validation import TypeValidator
 
 if TYPE_CHECKING:
     from oblate.contexts import LoadContext, DumpContext, ErrorContext
@@ -50,7 +51,6 @@ __all__ = (
 TD = TypeVar('TD', bound=TypedDictT)
 KT = TypeVar('KT')
 VT = TypeVar('VT')
-MISSING = utils.MISSING
 
 
 class _BaseStructField(Field[KT, VT]):
@@ -108,10 +108,13 @@ class Dict(_BaseStructField[DictT[KT, VT], DictT[KT, VT]]):
     _friendly_struct_name = 'dictionary'
 
     def __init__(self, key_tp: Type[KT] = MISSING, value_tp: Type[VT] = MISSING, /, **kwargs: Any) -> None:
+        self.key_tp = key_tp
+        self.value_tp = value_tp
+
         if key_tp is not MISSING:
             if value_tp is MISSING:
                 raise TypeError('Dict(T) is not valid, must provide a second argument for type of value')  # pragma: no cover
-            self._tp = utils.TypeValidator(DictT[key_tp, value_tp])
+            self._tp = TypeValidator({'root': DictT[key_tp, value_tp]})
         else:
             self._tp = None
 
@@ -121,7 +124,7 @@ class Dict(_BaseStructField[DictT[KT, VT], DictT[KT, VT]]):
         if not isinstance(value, dict):
             raise self._call_format_error(self.ERR_INVALID_DATATYPE, context.schema, value)
         if self._tp is not None:
-            validated, errors = self._tp.validate(value)
+            validated, errors = self._tp.validate('root', value)
             if not validated:
                 metadata = {'type_validation_fail_errors': errors}
                 raise self._call_format_error(self.ERR_TYPE_VALIDATION_FAILED, context.schema, value, metadata=metadata)
@@ -157,14 +160,15 @@ class TypedDict(_BaseStructField[TD, TD]):
     _friendly_struct_name = 'dictionary'
 
     def __init__(self, typed_dict: Type[TD], /, **kwargs: Any):
-        self._validator = utils.TypeValidator(typed_dict)
+        self.typed_dict = typed_dict
+        self._validator = TypeValidator({'root': typed_dict})
         super().__init__(**kwargs)
 
     def value_load(self, value: Any, context: LoadContext) -> TD:
         if not isinstance(value, dict):
             raise self._call_format_error(self.ERR_INVALID_DATATYPE, context.schema, value)
 
-        validated, errors = self._validator.validate(value)
+        validated, errors = self._validator.validate('root', value)
         if not validated:
             metadata = {'type_validation_fail_errors': errors}
             raise self._call_format_error(self.ERR_TYPE_VALIDATION_FAILED, context.schema, value, metadata=metadata)
@@ -191,22 +195,23 @@ class List(_BaseStructField[ListT[KT], ListT[KT]]):
 
     Parameters
     ----------
-    tp:
+    type:
         The type of list elements. This parameter corresponds to ``T``
         in ``typing.List[T]``.
     """
     _struct_name = 'list'
     _friendly_struct_name = 'list'
 
-    def __init__(self, tp: Type[KT] = MISSING, /, **kwargs: Any):
-        self._tp = None if tp is MISSING else utils.TypeValidator(ListT[tp])
+    def __init__(self, type: Type[KT] = MISSING, /, **kwargs: Any):
+        self.type = type
+        self._tp = None if type is MISSING else TypeValidator({'root': ListT[type]})
         super().__init__(**kwargs)
 
     def value_load(self, value: Any, context: LoadContext) -> ListT[KT]:
         if not isinstance(value, list):
             raise self._call_format_error(self.ERR_INVALID_DATATYPE, context.schema, value)
         if self._tp is not None:
-            validated, errors = self._tp.validate(value)
+            validated, errors = self._tp.validate('root', value)
             if not validated:
                 metadata = {'type_validation_fail_errors': errors}
                 raise self._call_format_error(self.ERR_TYPE_VALIDATION_FAILED, context.schema, value, metadata=metadata)
@@ -232,22 +237,23 @@ class Set(_BaseStructField[SetT[KT], SetT[KT]]):
 
     Parameters
     ----------
-    tp:
+    type:
         The type of set elements. This parameter corresponds to ``T``
         in ``typing.Set[T]``.
     """
     _struct_name = 'set'
     _friendly_struct_name = 'set'
 
-    def __init__(self, tp: Type[KT] = MISSING, /, **kwargs: Any):
-        self._tp = None if tp is MISSING else utils.TypeValidator(SetT[tp])
+    def __init__(self, type: Type[KT] = MISSING, /, **kwargs: Any):
+        self.type = type
+        self._tp = None if type is MISSING else TypeValidator({'root': SetT[type]})
         super().__init__(**kwargs)
 
     def value_load(self, value: Any, context: LoadContext) -> SetT[KT]:
         if not isinstance(value, set):
             raise self._call_format_error(self.ERR_INVALID_DATATYPE, context.schema, value)
         if self._tp is not None:
-            validated, errors = self._tp.validate(value)
+            validated, errors = self._tp.validate('root', value)
             if not validated:
                 metadata = {'type_validation_fail_errors': errors}
                 raise self._call_format_error(self.ERR_TYPE_VALIDATION_FAILED, context.schema, value, metadata=metadata)