+++ oblate/fields/typings.py
@@ -24,7 +24,7 @@ from __future__ import annotations
 
 from oblate.fields.base import Field
 from oblate.exceptions import FieldError
-from oblate import utils
+from oblate.type_validation import TypeValidator
 
 # typing imported as t to avoid name conflict with classes
 # defined in this module
@@ -42,6 +42,11 @@ __all__ = (
 
 _T = t.TypeVar('_T')
 
+def _generic_type_with_args(tp: _T, args: t.Sequence[t.Any]) -> _T:
+    # Before Python 3.11, T[*v] notation isn't supported and raises
+    # syntax error so this is a very hacky approach to get the desired
+    # generic type with given type arguments at runtime.
+    return tp.__getitem__(tuple(args))  # type: ignore
 
 class Any(Field[t.Any, t.Any]):
     """A field that accepts any arbitrary value.
@@ -49,8 +54,6 @@ class Any(Field[t.Any, t.Any]):
     This field acts as a "raw field" that performs no validation on the
     given value.
     """
-    __slots__ = ()
-
     def value_load(self, value: Any, context: LoadContext) -> Any:
         return value
 
@@ -75,28 +78,24 @@ class Literal(Field[_T, _T]):
     *values:
         The literal values.
     """
-    __slots__ = ('_values', '_msg')
-
     ERR_INVALID_VALUE = 'literal.invalid_value'
 
     def __init__(self, *values: _T, **kwargs: t.Any) -> None:
-        self._values = values
-        if len(values) == 1:
-            self._msg = f'Value must be {values[0]!r}'
-        else:
-            self._msg = f'Value must be one of: {", ".join(repr(value) for value in self._values)}'
+        self.values = values
+        self._tp = TypeValidator({'root': _generic_type_with_args(t.Literal, values)})
         super().__init__(**kwargs)
 
     def _get_default_error_message(self, error_code: t.Any, context: ErrorContext) -> t.Union[FieldError, str]:
         if error_code == self.ERR_INVALID_VALUE:
-            return self._msg
+            return FieldError(context.metadata['type_validation_fail_errors'])
 
         return super()._get_default_error_message(error_code, context)  # pragma: no cover
 
     def value_load(self, value: t.Any, context: LoadContext) -> _T:
-        if value not in self._values:
-            raise self._call_format_error(self.ERR_INVALID_VALUE, context.schema, value)
-
+        validated, errors = self._tp.validate('root', value)  # type: ignore
+        if not validated:
+            metadata = {'type_validation_fail_errors': errors}
+            raise self._call_format_error(self.ERR_INVALID_VALUE, context.schema, value, metadata)
         return value
 
     def value_dump(self, value: _T, context: DumpContext) -> t.Any:
@@ -121,15 +120,14 @@ class Union(Field[_T, _T]):
     *types: :class:`type`
         The list of types to accept.
     """
-    __slots__ = ('_tp',)
-
     ERR_INVALID_VALUE = 'union.invalid_value'
 
     def __init__(self, *types: t.Type[_T], **kwargs: t.Any):
         if len(types) < 2:
             raise TypeError('fields.Union() accepts at least two arguments')  # pragma: no cover
 
-        self._tp = utils.TypeValidator(t.Union[*types])  # type: ignore
+        self.types = types
+        self._tp = TypeValidator({'root': _generic_type_with_args(t.Union, types)})
         super().__init__(**kwargs)
 
     def _get_default_error_message(self, error_code: t.Any, context: ErrorContext) -> t.Union[FieldError, str]:
@@ -139,7 +137,7 @@ class Union(Field[_T, _T]):
         return super()._get_default_error_message(error_code, context)  # pragma: no cover
 
     def value_load(self, value: t.Any, context: LoadContext) -> _T:
-        validated, errors = self._tp.validate(value)  # type: ignore
+        validated, errors = self._tp.validate('root', value)  # type: ignore
         if not validated:
             metadata = {'type_validation_fail_errors': errors}
             raise self._call_format_error(self.ERR_INVALID_VALUE, context.schema, value, metadata)
@@ -155,22 +153,23 @@ class TypeExpr(Field[_T, _T]):
     For the list of supported types and limitations of this field, please see
     the :ref:`guide-type-validation` section.
 
+    .. note::
+
+        This uses :func:`oblate.validate_types` under the hood.
+
     .. versionadded:: 1.1
 
     Parameters
     ----------
-    tp:
+    expr:
         The type expression that should be used to validate the type of
         given value.
     """
-    __slots__ = (
-        '_tp',
-    )
-
     ERR_TYPE_VALIDATION_FAILED = 'type_expr.type_validation_failed'
 
-    def __init__(self, tp: t.Type[_T], **kwargs: t.Any):
-        self._tp = utils.TypeValidator(tp)
+    def __init__(self, expr: t.Type[_T], **kwargs: t.Any):
+        self.expr = expr
+        self._tp = TypeValidator({'root': expr})
         super().__init__(**kwargs)
 
     def _get_default_error_message(self, error_code: t.Any, context: ErrorContext) -> t.Union[FieldError, str]:
@@ -180,7 +179,7 @@ class TypeExpr(Field[_T, _T]):
         return super()._get_default_error_message(error_code, context)  # pragma: no cover
 
     def value_load(self, value: t.Any, context: LoadContext) -> _T:
-        validated, errors = self._tp.validate(value)
+        validated, errors = self._tp.validate('root', value)
         if not validated:
             metadata = {'type_validation_fail_errors': errors}
             raise self._call_format_error(self.ERR_TYPE_VALIDATION_FAILED, context.schema, value, metadata)